+++ packages/hooks/use-in-view/__tests__/use-in-view.test.tsx
@@ -1,8 +1,8 @@
 import ReactDOM from "react-dom";
 import { act } from "react-dom/test-utils";
-import * as useInView from "../use-in-view";
+import * as useInView from "../";
 
-let container;
+let container: HTMLDivElement;
 
 beforeEach(() => {
   container = document.createElement("div");